+++ src/main/java/com/getpcpanel/ui/MiniLightingDialog.java
@@ -105,6 +105,7 @@ public class MiniLightingDialog extends Application implements UIInitializer<Sin
     @Override
     public void start(Stage stage) {
         this.stage = stage;
+        UIHelper.closeOnEscape(stage);
         var scene = new Scene(Objects.requireNonNull(root));
         scene.getStylesheets().add(Objects.requireNonNull(getClass().getResource("/assets/dark_theme.css")).toExternalForm());
         stage.getIcons().add(new Image(Objects.requireNonNull(getClass().getResource("/assets/256x256.png")).toExternalForm()));