+++ src/main/java/com/getpcpanel/ui/ProLightingDialog.java
@@ -144,6 +144,7 @@ public class ProLightingDialog extends Application implements UIInitializer<Sing
     @Override
     public void start(Stage stage) {
         this.stage = stage;
+        UIHelper.closeOnEscape(stage);
         var scene = new Scene(root);
         scene.getStylesheets().add(Objects.requireNonNull(getClass().getResource("/assets/dark_theme.css")).toExternalForm());
         stage.getIcons().add(new Image(Objects.requireNonNull(getClass().getResource("/assets/256x256.png")).toExternalForm()));