+++ src/main/java/com/getpcpanel/ui/RGBLightingDialog.java
@@ -88,6 +88,7 @@ public class RGBLightingDialog extends Application implements UIInitializer<Sing
     @Override
     public void start(Stage stage) {
         this.stage = stage;
+        UIHelper.closeOnEscape(stage);
         var scene = new Scene(root);
         scene.getStylesheets().add(Objects.requireNonNull(getClass().getResource("/assets/dark_theme.css")).toExternalForm());
         stage.getIcons().add(new Image(Objects.requireNonNull(getClass().getResource("/assets/256x256.png")).toExternalForm()));